Tweaks And New Base Pokémon Data
Tweaks to the Pokemon related code and the GBP music engine (collecting
all the flags into a single status flag bitfield).

Overhauling Pokémon evolution and base stats data and related code to
simplify the accessing of code.

Note that the base data for the Pokémon is currently still being
converted to the new format and is therefore absent from this commit.