Initial Implementation Overworld Object Management
Began implementing the overworld object management software to track the
on screen entities.

Removed templating of LinkedList class to simplify generated code and to
optimise repeated code out.

Re-added stereo playing functionality to the GB Player music engine.

Added Initialisation to the Game class to initialise the options to
sensible values for testing.