Palette Fading Tweaks & Saving Functionality
Tweaks to the palette fading code to support exit and enter callbacks on
ending the fade. This allows code for the GameScreens to be executed
when they leave a screen (best used to select the new screen to
transition to) or when they become active after fading in.

Partial conversion of M4A Engine to new format.

Addition of Flash 128k save and load code.

Addition of RTC palette updates for objects, backgrounds yet to be
implemented.

Saving handled by passing a pointer to a SaveLocationStruct to the
FlashFunctions' WriteToFlash or ReadFromFlash methods. The code will
then appropriately copy data out from the Flash save ram, or to the ram.
System is easily extensible simply by adding your own structure array
which specifies the data which requires saving.

0xFFC bytes in a block, with the last 4 bytes reserved for the checksum.
Note that the code does not check for blocks being too large or
overflowing into other sectors, so you will have to manage them
yourself.