Switched out a large number of previous calbacks systems which made use of function pointers and replaced them with the vastly more OO compliant Callback classes.
Split the fade functions and related Palette functionality into its own class. Updates before Game, InputManager and SceneManager to ensure that any scene changes or other related callbacks caused by fades execute at the start of the frame.

Implemented the NewGameScreen and LoadGameScreen, both of which are selectively accessed from the title screen and options screen. They allow you to start and load games on the engine, as well as offering pre-game start access to the options menu.

Implemented the options menu. Can currently be accessed from the NewGame and LoadGame Screens. In future, will need to be accessible from the main in game menu too.

Complete rework of the ToString function belonging to the String class. As a result, leading zeroes and formatting is easier to choose and the method works far more reliably.

Placeholder implementation of the WelcomeToTheWorldOfPokemon screen - the screen you first see upon starting a new game.