Added saving functionality - note however, that is is buggy and non-functional as of this commit. It does not, however, crash the game, merely locks input because the script currently waits until the save has completed, which it never does.
Added in the effects of lock, lockall, release and releaseall to the scripting engine. As a result, scripts no longer refuse to release the player on ending if the release or releaseall commands have been used. lockall is used by default during saving to lock all NPCs in place.

Updated the IRQHandler's Prime function to allow you to attach a function pointer as an argument to the IRQ being used. By default, it is null.

Introduced the concept of Asynchronous tasks to the engine. Unlike threading, which is truly async, these still rely on synchronous updates along with the rest of the engine. However, they allow the task being handled to be split into separate and disparate parts which may be executed over the course of an arbitrarily large number of frames (to the limit of unsigned 32 bit numbers). As an example, saving is an async task, executed over the course of a number of frames. This is done to prevent hanging, in theory. Currently, the engine still hangs while saving, meaning the saving task has not been subdivided enough as yet.

Changes and tweaks to built in scripts to correct small errors and oversights.

Overhaul of pointers in the engine to use Smart Pointers instead of standard pointers. Introduces a level of overhead, but is vastly safer to use at runtime.